LayoutTests/editing/execCommand/query-font-size-expected.txt

test	html	queryCommandValue result	getComputedStyle result
execCommand('FontSize', -2)	<span class="Apple-style-span" style="font-size: x-small;">test</span>	1	10px
execCommand('FontSize', -1)	<span class="Apple-style-span" style="font-size: small;">test</span>	2	13px
execCommand('FontSize', 0)	<span>test</span>	3	16px
execCommand('FontSize', 1)	<span class="Apple-style-span" style="font-size: x-small;">test</span>	1	10px
execCommand('FontSize', 2)	<span class="Apple-style-span" style="font-size: small;">test</span>	2	13px
execCommand('FontSize', 3)	<span>test</span>	3	16px
execCommand('FontSize', 4)	<span class="Apple-style-span" style="font-size: large;">test</span>	4	18px
execCommand('FontSize', 5)	<span class="Apple-style-span" style="font-size: x-large;">test</span>	5	24px
execCommand('FontSize', 6)	<span class="Apple-style-span" style="font-size: xx-large;">test</span>	6	32px
execCommand('FontSize', 7)	<span class="Apple-style-span" style="font-size: -webkit-xxx-large;">test</span>	7	48px
execCommand('FontSize', '8px')	<span class="Apple-style-span" style="font-size: -webkit-xxx-large;">test</span>	7	48px
execCommand('FontSize', '2px')	<span class="Apple-style-span" style="font-size: small;">test</span>	2	13px
manual CSS font-size: 3px	<span style="font-size: 3px">test</span>	1	3px
manual CSS font-size: 0.2em	<span style="font-size: 0.2em">test</span>	1	9px
manual CSS font-size: 17px	<span style="font-size: 17px">test</span>	4	17px
manual CSS font-size: 31px	<span style="font-size: 31px">test</span>	6	31px
manual CSS font-size: 50px	<span style="font-size: 50px">test</span>	7	50px
manual CSS font-size: 5em	<span style="font-size: 5em">test</span>	7	80px
manual CSS font-size: 10px	<span style="font-size: 10px">test</span>	1	10px
monospace tests to show bug 19161
execCommand('FontSize', -2)	<span class="Apple-style-span" style="font-size: x-small;"><font class="Apple-style-span" face="monospace">test</font></span>	1	10px
execCommand('FontSize', -1)	<span class="Apple-style-span" style="font-size: small;"><font class="Apple-style-span" face="monospace">test</font></span>	2	13px
execCommand('FontSize', 0)	<span><font class="Apple-style-span" face="monospace">test</font></span>	3	16px
execCommand('FontSize', 1)	<span class="Apple-style-span" style="font-size: x-small;"><font class="Apple-style-span" face="monospace">test</font></span>	1	10px
execCommand('FontSize', 2)	<span class="Apple-style-span" style="font-size: small;"><font class="Apple-style-span" face="monospace">test</font></span>	2	13px
execCommand('FontSize', 3)	<span><font class="Apple-style-span" face="monospace">test</font></span>	3	16px
execCommand('FontSize', 4)	<span class="Apple-style-span" style="font-size: large;"><font class="Apple-style-span" face="monospace">test</font></span>	4	18px
execCommand('FontSize', 5)	<span class="Apple-style-span" style="font-size: x-large;"><font class="Apple-style-span" face="monospace">test</font></span>	5	24px
execCommand('FontSize', 6)	<span class="Apple-style-span" style="font-size: xx-large;"><font class="Apple-style-span" face="monospace">test</font></span>	6	32px
execCommand('FontSize', 7)	<span class="Apple-style-span" style="font-size: -webkit-xxx-large;"><font class="Apple-style-span" face="monospace">test</font></span>	7	48px
